test: skip the specific test for future fix
This test checks the KUMO rewards token to the users of stability pools when a new user joins the SP. Removing KUMO tokens work is in progress and there is a possibility that this test will fail by the updates. Later It will be updated according to the changes by the backend.