chore(sol): remove redundant line
This line was meant to remove a front-end tag and make a deposit equal to zero, because `deposits` was a struct with two variables. But after removing front-end tags, and with the if condition in the line above, the deposit value is already equal to zero to that moment. So, this line is redundant.