fix: fixed build, failed package dependency in subgraph
Build was failing because of one the nested dependency in the subgraph
workspace package and gettting error
Trace:
Error: https://codeload.github.com/hugomrdias/concat-stream/tar.gz/....:
Request failed "404 Not Found"
This package is the dependency of  ipfs-http-client@^34.0.0:
The  ipfs-http-client is the dependency of @graphprotocol/graph-cli@^0.20.0
Although it is prefixes with ^ to update for both minor and patch update but
not working as expected and getting other types (typescript) errors.
Updated it ~ to update the latest patch only and then it works.
To keep the other packages unaffected updated the resolution field in
package.json of root dir with "devDependencies" of dev-frontend package.