build(sol): update to latest sol version
In line with security best practices, all pragma statements
are changed to the latest stable solidity version 0.8.15.
This is latest version supported by the currently used hardhat version.

The statement omits the caret and tilde so that there is less
ambiguity about the version in which the code was compiled.